feat(the-rust/cpc): restructure cpc package and add backyard example package
