+++ the-rust-programming-language/common-programming-concepts/src/data_types_and_control_flows.rs
@@ -1,0 +1,11 @@
+pub mod variables;
+pub mod shadowing;
+pub mod numeric_operations;
+pub mod functions;
+pub mod the_slice_type;
+pub mod struct_method;
+mod struct_method_multiple_impl;
+mod struct_method_multiple_parameters;
+pub mod structs;
+pub mod enums;
+pub mod loops;
