+++ the-rust-programming-language/common-programming-concepts/src/data_types_and_control_flows/struct_method_multiple_parameters.rs
