+++ the-rust-programming-language/common-programming-concepts/src/data_types_and_control_flows/structs.rs
@@ -30,7 +30,7 @@ fn print_user_struct() {
     println!("email: {}", user.email);
 
     let user2 = User {
-        email: String::from("another@example.com"),
+        email: String::from("another@ownership_and_references.com"),
         ..user
     };
