+++ the-rust-programming-language/common-programming-concepts/src/ownership_and_references.rs
@@ -1,0 +1,5 @@
+pub mod dangling_references;
+pub mod reference_and_borrowing;
+pub mod ownership_and_function;
+pub mod return_values_and_scope;
+pub mod scopes;
