+++ the-rust-programming-language/common-programming-concepts/src/ownership_and_references/dangling_references.rs
