+++ the-rust-programming-language/common-programming-concepts/src/ownership_and_references/ownership_and_function.rs
