+++ the-rust-programming-language/common-programming-concepts/src/ownership_and_references/return_values_and_scope.rs
